chore(core): moved linux FS detect to linux-specific file
